pyproject.toml
[tool.poetry]
name = "viking-log-keeper"
version = "0.7.0"
description = "661 VGS - Function to collate 2965D log sheets into a master log and database."
authors = ["Michael Jennings <mjennings061@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
packages = [
    { include = "log_keeper", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.4"
pymongo = {extras = ["srv"], version = "^4.6.1"}
xlsxwriter = "^3.1.9"
cryptography = "^41.0.7"
python-dotenv = "^1.0.0"
inquirer = "^3.1.4"
openpyxl = "^3.1.2"

[tool.poetry.urls]
homepage = "https://github.com/mjennings061/viking-log-keeper"
repository = "https://github.com/mjennings061/viking-log-keeper.git"

[tool.poetry.scripts]
update-logs = "log_keeper.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
